V 1.1

+++ src/main/java/com/example/Spring_project/dto/BalanceResponse.java
@@ -1,14 +1,15 @@
 package com.example.Spring_project.dto;
 
+import java.math.BigDecimal;
+
 public class BalanceResponse {
     private String phoneNumber;
-    private int balance;
+    private BigDecimal balance;
 
     public void setPhoneNumber(String newNumberPhone) {
         phoneNumber = newNumberPhone;
     }
-
-    public void setBalance(int newBalance) {
+    public void setBalance(BigDecimal newBalance) {
         balance = newBalance;
     }
 }
