+++ src/main/java/com/example/Spring_project/service/UserService.java
@@ -7,32 +7,33 @@ import com.example.Spring_project.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import java.math.BigDecimal;
 
 @Service
 public class UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) { // Конструктор
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
     }
 
     @Transactional
     public User registerUser(AuthRequest authRequest) { // Регистрация пользователя
-        if (userRepository.existsByPhoneNumber(authRequest.getPhoneNumber())) {
+        if (userRepository.existsByPhoneNumber(authRequest.getPhoneNumber())) { // Если пользователь уже сушествует
             throw new RuntimeException("Пользователь уже существует");
         }
 
         User user = new User();
         user.setPhoneNumber(authRequest.getPhoneNumber());
         user.setPassword(passwordEncoder.encode(authRequest.getPassword()));
-        user.setBalance(1000); // Начальный баланс 1000 руб
+        user.setBalance(BigDecimal.valueOf(1000.00)); // Начальный баланс 1000 руб
 
         return userRepository.save(user);
     }
 
-    public BalanceResponse getBalance(User user) { // Получение баланса
+    public BalanceResponse getBalance(User user) {
         BalanceResponse response = new BalanceResponse();
         response.setPhoneNumber(user.getPhoneNumber());
         response.setBalance(user.getBalance());
